pyproject.toml
[tool.poetry]
name = "furaffinity_scrape"
version = "0.1.0"
description = "scrape utilites for furaffinity.net"
authors = ["Mark Grandi <markgrandi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
arrow = "^0.17.0"
attrs = "^20.3.0"
logging_tree = "^1.8.1"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.2"
pyhocon = "^0.3.57"

[tool.poetry.dev-dependencies]
wheel = "^0.36.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
